+++ lib/messaging/messagehandler_status.rb
@@ -1,35 +1,14 @@
 require 'json'
 require './lib/database/dbaccess.rb'
 require 'time'
+require_relative 'messagehandler_base'
 
 # Get the JSON command, received through skynet, and send it to the farmbot
 # command queue Parses JSON messages received through SkyNet.
-class MessageHandlerStatus
+class MessageHandlerStatus < MessageHandlerBase
 
-  attr_accessor :message
-
-  ## general handling messages
-
-  def initialize
-    @dbaccess = $bot_dbaccess
-    @last_time_stamp  = ''
-  end
-
-  # A list of MessageHandler methods (as strings) that a Skynet User may access.
-  #
   def whitelist
     ["read_status"]
-  end
-
-  # Handle the message received from skynet
-  #
-  def handle_message(message)
-
-    if whitelist.include?(message.message_type)
-      self.send(message.message_type, message)
-      message.handled = true
-    end
-
   end
 
   # Send the current status to the requester
